// src/app/portal/legalNatures/[id]/schema.ts
import { z } from 'zod';

export const schemaLegalNature = z.object({
    id: z.number().optional(),
    code: z.string().optional(),
    name: z.string().optional(),
    slug: z.string().optional(),
    active: z.boolean().optional(),
    dtinsert: z.date().optional(),
    dtupdate: z.date().optional(),
});